+++ apps/fe/src/routes/_main/learning/question.tsx
@@ -1,69 +1,15 @@
-import { useCallback } from 'react';
-
-import { getFeedbackApi, submitRecordApi } from '@/apis/learning-api';
 import AnswerSection from '@/features/learning/components/question/answer-section';
 import FeedbackSection from '@/features/learning/components/question/feedback-section';
 import FloatingStepBar from '@/features/learning/components/question/floating-step-bar';
 import QuestionContent from '@/features/learning/components/question/question-content';
 import QuestionHeader from '@/features/learning/components/question/question-header';
 import VoiceRecorderSection from '@/features/learning/components/question/voice-recorder-section';
-import { useAssessmentStream } from '@/features/learning/lib/hooks/use-assessment-stream';
-import useLearningSession, { ANSWER_PHASE } from '@/lib/stores/learning-session';
+import useLearningSession from '@/lib/stores/learning-session';
 import { createFileRoute, useNavigate } from '@tanstack/react-router';
 
 const QuestionPage = () => {
-  const sessionId = useLearningSession((state) => state.sessionId);
   const question = useLearningSession((state) => state.question);
-  const setRemainedCredit = useLearningSession((state) => state.setRemainedCredit);
-  const answerId = useLearningSession((state) => state.answer.answerId);
-  const setPhase = useLearningSession((state) => state.setPhase);
-  const setSttText = useLearningSession((state) => state.setSttText);
-  const setFeedback = useLearningSession((state) => state.setFeedback);
-  const setAssessmentStatus = useLearningSession((state) => state.setAssessmentStatus);
-
   const navigate = useNavigate();
-
-  const handleFeedbackDone = useCallback(async () => {
-    if (!answerId) return;
-
-    try {
-      const feedback = await getFeedbackApi(answerId);
-      setFeedback(feedback);
-      setRemainedCredit(feedback.remainingToken);
-      setPhase(ANSWER_PHASE.FEEDBACK_DONE);
-    } catch (error) {
-      console.error('피드백 조회 실패:', error);
-    }
-  }, [answerId, setFeedback, setPhase, setRemainedCredit]);
-
-  useAssessmentStream({
-    sessionId,
-    answerId,
-    onStatusChange: setAssessmentStatus,
-    onDone: handleFeedbackDone,
-  });
-
-  const handleRecordingComplete = async (audioBlob: Blob) => {
-    if (!sessionId || !question) return;
-
-    try {
-      setPhase(ANSWER_PHASE.STT_LOADING);
-
-      const extension = audioBlob.type.split('/')[1]?.split(';')[0] || 'webm';
-      const audioFile = new File([audioBlob], `answer.${extension}`, { type: audioBlob.type });
-      const { sttText } = await submitRecordApi({
-        sessionId,
-        questionId: question.questionId,
-        extraQuestionId: question.extraQuestionId,
-        audioFile,
-      });
-
-      setSttText(sttText);
-      setPhase(ANSWER_PHASE.STT_DONE);
-    } catch (error) {
-      console.error('녹음 제출 실패:', error);
-    }
-  };
 
   if (!question) {
     navigate({ to: '/learning', replace: true });
@@ -78,10 +24,7 @@ const QuestionPage = () => {
     <div className="relative mx-auto flex min-h-screen max-w-250 flex-col gap-8 p-6 sm:p-10">
       <QuestionHeader />
       <QuestionContent key={`question-content-${sessionKey}`} />
-      <VoiceRecorderSection
-        key={`voice-recorder-section-${sessionKey}`}
-        onRecordingComplete={handleRecordingComplete}
-      />
+      <VoiceRecorderSection key={`voice-recorder-section-${sessionKey}`} />
       <AnswerSection />
       <FeedbackSection />
       <div className="flex-1" />
